remove flaky label from eviction tests
